Additional firemode file split
After how much the Universal Remote's grown, split main firemodes file into two additional sub-files. firemodes_utility for firemodes that provide buffs or other non-combat options, and
firemodes_piles specifically for cluster prop breakup and "supply crate" pickup functionality
